Refactored ManageSchedule into smaller components to improve code maintainance and scalability.
